feat: [FE] Tournament typescript 분리
