--- src/types.ts
@@ -1,5 +1,0 @@
-export interface GameDTO {
-    r_name: string;
-    r_img: string;
-    idx: number;
-}
